Restructured the modules to avoid generic imports.
Also made sure pep8 is respected.
