+++ statsmodels/kernel_methods/api.py
@@ -1,11 +1,12 @@
-__all__ = [
-    'kde', 'KDE', 'bandwidths', 'kde_methods', 'kernels', 'Grid', 'kde_utils'
-]
-
 from . import kde
 from .kde import KDE
 from . import bandwidths
-from . import kde_methods
-from . import kernels
+from . import kde_1d, kde_nd, kde_nc, kde_multivariate
+from . import kernels1d, kernelsnc, kernelsnd
 from .kde_utils import Grid
 from . import kde_utils
+
+__all__ = [
+    'kde', 'KDE', 'bandwidths', 'kde_1d', 'kde_nd', 'kde_nc', 'kernels1d',
+    'kernelsnc', 'kernelsnd', 'Grid', 'kde_utils'
+]
