+++ statsmodels/kernel_methods/kernelsnc.py
